chore(release): 03/02/20

+++ protocols/zymo-ribofree-p5-adapter-ligation/3_p5_adapter_ligation.ot2.apiv2.py
@@ -18,20 +18,20 @@ def run(ctx):
     # [number_of_samples, p20_mount, p50_mount] = [96, 'right', 'left']
 
     # load modules and labware
-    tempdeck = ctx.load_module('tempdeck', '1')
-    tempdeck.set_temperature(4)
-    tempblock = tempdeck.load_labware(
-        'opentrons_24_aluminumblock_nest_1.5ml_screwcap')
     tc = ctx.load_module('thermocycler')
     tc.set_lid_temperature(100)
     tc.set_block_temperature(4)
     tc_plate = tc.load_labware('nest_96_wellplate_100ul_pcr_full_skirt')
-    reagent_res = ctx.load_labware(
-        'nest_12_reservoir_15ml', '2', 'reagent reservoir')
     racks20 = [
         ctx.load_labware('opentrons_96_tiprack_20ul', slot)
-        for slot in ['3', '4', '5', '6']
+        for slot in ['1', '2', '3', '6']
     ]
+    tempdeck = ctx.load_module('tempdeck', '4')
+    tempdeck.set_temperature(4)
+    tempblock = tempdeck.load_labware(
+        'opentrons_24_aluminumblock_nest_1.5ml_screwcap')
+    reagent_res = ctx.load_labware(
+        'nest_12_reservoir_15ml', '5', 'reagent reservoir')
     racks50 = [ctx.load_labware('opentrons_96_tiprack_300ul', '9')]
 
     # pipettes
@@ -43,19 +43,20 @@ def run(ctx):
     p20.flow_rate.blow_out = 30
     m50 = ctx.load_instrument('p50_multi', p50_mount, tip_racks=racks50)
 
-    file_path = '/data/csv/tip_track.json'
-    # file_path = 'protocols/tip_track.json'
-    if os.path.isfile(file_path):
-        with open(file_path) as json_file:
-            data = json.load(json_file)
-            if 'tips20' in data:
-                tip20_count = data['tips20']
-            else:
-                tip20_count = 0
-            if 'tips50' in data:
-                tip50_count = data['tips50']
-            else:
-                tip50_count = 0
+    if not ctx.is_simulating():
+        file_path = '/data/csv/tip_track.json'
+        # file_path = 'protocols/tip_track.json'
+        if os.path.isfile(file_path):
+            with open(file_path) as json_file:
+                data = json.load(json_file)
+                if 'tips20' in data:
+                    tip20_count = data['tips20']
+                else:
+                    tip20_count = 0
+                if 'tips50' in data:
+                    tip50_count = data['tips50']
+                else:
+                    tip50_count = 0
     else:
         tip20_count = 0
         tip50_count = 0
@@ -126,9 +127,10 @@ with cleanup.')
 
     # track final used tip
     # file_path = '/data/csv/tip_track.json'
-    data = {
-        'tips20': tip20_count,
-        'tips50': tip50_count
-    }
-    with open(file_path, 'w') as outfile:
-        json.dump(data, outfile)
+    if not ctx.is_simulating():
+        data = {
+            'tips20': tip20_count,
+            'tips50': tip50_count
+        }
+        with open(file_path, 'w') as outfile:
+            json.dump(data, outfile)
